requirements.txt
streamlit
torch
torchvision
pandas
numpy
scikit-learn
joblib
matplotlib
Pillow
